This PR does make changes to the database schema. It won't break existing databases, but the Fix for #7 will only work once the summoner and match tables have been rebuild or the affected columns have been changed manually to BigInteger:

match_participant_identities.p_accountId
match_participant_identities.p_currentAccountId
summoner.accountId
Also includes a little bugfix for edge cases in constants.

+++ cassiopeia-sqlstore/cassiopeia_sqlstore/match.py
@@ -201,10 +201,10 @@ class SQLMatchParticipantsIdentities(SQLBaseObject):
                     Column("p_platformId", String(5)),
                     Column("match_gameId", BigInteger, primary_key=True),
                     Column("participantId", Integer, primary_key=True),
-                    Column("p_accountId", Integer),
+                    Column("p_accountId", BigInteger),
                     Column("p_summonerName", String),
                     Column("p_summonerId", Integer),                    
-                    Column("p_currentAccountId", Integer),
+                    Column("p_currentAccountId", BigInteger),
                     Column("p_profileIcon", Integer),
                     ForeignKeyConstraint(
                         ["p_currentPlatformId","match_gameId"],
